Logic updated.
Modularization of responsibilities.
