+++ README.md
@@ -1,5 +1,4 @@
 # Juego de Damas
-
 Este es un juego de damas en Python donde un jugador puede jugar contra la computadora.
 
 ## Instrucciones para ejecutar el proyecto
@@ -7,14 +6,18 @@ Este es un juego de damas en Python donde un jugador puede jugar contra la compu
 2. **Ejecutar juego**
 3. **Otro**
 
+## Instrucciones para testing:
+1. **Ejecutar: ....**
+
 ## Relgas del juego
-- Las fichas negras comienzan en la parte superior y las fichas blancas en la parte inferior.
-- Las fichas se mueven en diagonal.
+- Las fichas negras comienzan en la parte inferior y las fichas rojas en la parte superior.
+- Las fichas se mueven únicamente avanzando en diagonal.
 - Puedes capturar fichas del oponente saltando sobre ellas.
 
 ## Estructura del Proyecto
-- Damas.py: Contiene el código principal del juego.
+- Damas.py: Inicia la interfaz gráfica del juego.
+- logica.py: Contiene la lógica inicial del juego.
 
 ## Futuras mejoras
 - Implementar la lógica de la IA para la computadora.
-- Agregar una interfaz gráfica.
+- Agregar una interfaz gráfica mejorada.
